app-server: Split transport module
